+++ confignode/src/main/java/org/apache/iotdb/confignode/manager/load/LoadManager.java
@@ -88,8 +88,7 @@ public class LoadManager {
         new StatisticsService(configManager, routeBalancer, loadCache, loadPublisher);
 
     loadPublisher.register(statisticsService);
-    // TODO: enable
-    // loadPublisher.register(configManager.getPipeManager().getPipeRuntimeCoordinator());
+    loadPublisher.register(configManager.getPipeManager().getPipeRuntimeCoordinator());
   }
 
   /**
